Add geom option to remove zdc (#879)
### Briefly, what does this PR introduce?

This PR adds an additional geometry compact file which removes the ZDC
for specialized running where ZDC is not required. This allows for e+A
simulations to run about 10x faster (and produce smaller output files)
due to the enormous overhead necessary for ZDC simulations (high energy
neutrons into a 2m long HCAL).

this is in response to https://github.com/eic/EICrecon/issues/1888 

### What kind of change does this PR introduce?
- [ ] Bug fix (issue #__)
- [x ] New feature (issue (https://github.com/eic/EICrecon/issues/1888))
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [x ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [x ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?

No, but anyone using this compact file will NOT have ZDC information in
the output.

### Does this PR change default behavior?

Only for people using this compact file.

---------

Co-authored-by: Alexander Jentsch <ajentsch@bnl.gov>
Co-authored-by: Dmitry Kalinkin <dmitry.kalinkin@gmail.com>