feat: make FieldMapB rotation and translation optional (#895)
### Briefly, what does this PR introduce?
This PR makes the specification of rotation and translation for
FieldMapB entities optional, and does not do any trivial rotation and
translation when they are not set. VTune profiling indicated that there
was a 1% speed-up opportunity in avoiding the ROOT::Math::XYZPoint
transformations (as well as the delayed phi determination). It is a bit
of a micro-optimization, though, but `fieldComponents` is still the
highest nail that is entirely under our control in simulations (because
poorly defined geometry doesn't show up that way...).

### What kind of change does this PR introduce?
- [ ] Bug fix (issue #__)
- [x] New feature (issue: avoid defining trivial objects)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No.

### Does this PR change default behavior?
No.