Add support materials to FTOF (#925)
### Briefly, what does this PR introduce?

There is currently nothing between the front and back layers of FTOF.
However, dRICH group wants to study the effect of FTOF material and
having the right material is important for their study.

A uniform slab of carbon foam of width 1 cm is now sandwiched between
the two layers. Details such as cooling tubes will be added later. The
width of the support structure is given by slide 6 of
https://indico.bnl.gov/event/24949/attachments/57520/98759/ePICTOF_WP2_FTOFLayout_09272024v2.pdf

The thickness of the module board and PCB are not final and will be
changed in the future.

### What kind of change does this PR introduce?
- [ ] Bug fix (issue #__)
- [x] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?

No

### Does this PR change default behavior?

No

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Dmitry Kalinkin <dmitry.kalinkin@gmail.com>