ForwardServiceGap_length increased to move endcap calorimeters back by 2.8cm
The project envelope (9-22-25, and previous iterations) has the back of the LFHCAL at 4965mm, 28mm farther than the current simulation geometry.  The easiest way to move all of the P calorimeter's backward is to increase the gap.  

That said, I would recommend using the project envelope z_max as the anchor point, and run the calculation outside-in, so changes in the tracker and PID volumes don't move the forward endcap in the future.  However, Sakib asked me not to implement that, so I've tried the current approach.