Flat ob (#933)
### Briefly, what does this PR introduce?

This is a new version of the outer barrel geometry silicon_barrel.xml to
better match the CAD files. Changes include;

Replacing the carbon fiber frame with all flat components
Adding carbon fiber braces to model peaks in the material thickness due
to module structure
Adding gap in the active silicon in the centre of modules to model the
dead area
Replacing the tilted stave alignment with a castellated arrangement
Small changes to widths and thicknesses

This version matches the material thickness of CAD model to ~5% and is
significantly less than the current epic-main geometry

Work done by Aditya Yalavatti and Sam Henry, University of Oxford

### What kind of change does this PR introduce?
- [ ] Bug fix (issue #__)
- [X ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?

### Does this PR change default behavior?

---------

Co-authored-by: Samuel Henry <henry@pplxint13.physics.ox.ac.uk>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Samuel Henry <henry@pplxint12.physics.ox.ac.uk>
Co-authored-by: Athavan Ramalingam <Gitavan@users.noreply.github.com>
Co-authored-by: Dmitry Kalinkin <dmitry.kalinkin@gmail.com>
Co-authored-by: Shujie Li <shujie@jlab.org>
Co-authored-by: ShujieL <ShujieL@users.noreply.github.com>