ForwardServiceGap_length increased by 2.8cm to move calo endcap (#960)
### Briefly, what does this PR introduce?

The project envelope (9-22-25, and previous iterations) has the back of
the LFHCAL at 4965mm, 27.6 mm farther than the current simulation
geometry (which ends at z=4937.4 mm) . The easiest way to move all of
the P calorimeters backward is to increase the ForwardServiceGap_length
by 2.76cm.

That said, I would recommend using the project envelope z_max as the
anchor point, and run the calculation outside-in, so changes in the
tracker and PID volumes don't move the forward endcap in the future.
However, Sakib asked me not to implement that at first, so I've tried
the current approach (but I would prefer this).

For a bit more detail, the current front position of the LFHCAL is based
on these values:

LFHCAL_zmin = EcalEndcapP_zmin + EcalEndcapP_length
      = ForwardServiceGap_zmax + 30*cm
      = ForwardServiceGap_zmin + ForwardServiceGap_length + 30*cm
= ForwardPIDRegion_zmin + ForwardInnerEndcapRegion_length +
ForwardInnerEndcapRegionExtraSpace_length + 9.6*cm + 30*cm
= CentralTrackingRegionP_zmax + ForwardPIDRegion_length +
ForwardTrackingRegion_length + 4.6*cm + 9.6*cm + 30*cm
= 1800.0*mm + 135.0*cm + 0.0*cm + 4.6*cm + 9.6*cm + 30*cm  = 3,592 mm

which puts the LFHCAL z_max at 4937.4mm, again 27.6mm closer to the IP
than specified in the project envelope.

### What kind of change does this PR introduce?
- [ ] Bug fix (issue #__)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ X] Other: __ Compliance with project envelope (9-22-25)

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [X ] Documentation has been added / updated
- [X ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?

None - it just moves the forward calorimeters back by 2.8cm

### Does this PR change default behavior?

It will change, very slightly, the acceptance of the forward
calorimeters.

---------

Co-authored-by: Dmitry Kalinkin <dmitry.kalinkin@gmail.com>