Separate tagger from beampipe (#968)
### Briefly, what does this PR introduce?
Separates the definition of the Tagger detector from the beamline
vacuum. The taggers are now inserted into the world volume rather than
added into a section of the beampipe which is split between vacuum and
air.

There are some very small changes to the positions of the detectors to
match small edits to the beamline vacuum.

### What kind of change does this PR introduce?
- [x] Bug fix (issue #963)
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?
No
### Does this PR change default behavior?
No

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Dmitry Kalinkin <dmitry.kalinkin@gmail.com>