Allow virtual beampipe tracking detector volume to be intersection solid (#989)
### Briefly, what does this PR introduce?
In preparation for changes in the shape of the beampipe in the far
backward region this PR allows the tracking layers to be defined as
intersection solids between any shaped beampipe and a thin box.

The locations where it becomes important to track the position of the
beamspot and acceptance of Low-Q2 electrons along the beamline is
expected to move away from the current cone segment. This prepares for
that change by ensuring that the current benchmark plots will remain
valid.

### What kind of change does this PR introduce?
- [x] Bug fix (issue #963 )
- [ ] New feature (issue #__)
- [ ] Documentation update
- [ ] Other: __

### Please check if this PR fulfills the following:
- [ ] Tests for the changes have been added
- [ ] Documentation has been added / updated
- [ ] Changes have been communicated to collaborators

### Does this PR introduce breaking changes? What changes might users
need to make to their code?

### Does this PR change default behavior?

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>